Use getJSXOpeningElementName utility
