+++ src/rules/direct-slot-children.js
@@ -1,4 +1,5 @@
 const {isPrimerComponent} = require('../utils/is-primer-component')
+const {getJSXOpeningElementName} = require('../utils/get-jsx-opening-element-name')
 const {last} = require('lodash')
 
 const slotParentToChildMap = {
@@ -82,14 +83,5 @@ module.exports = {
         stack.pop()
       }
     }
-  }
-}
-
-// Convert JSXOpeningElement name to string
-function getJSXOpeningElementName(jsxNode) {
-  if (jsxNode.name.type === 'JSXIdentifier') {
-    return jsxNode.name.name
-  } else if (jsxNode.name.type === 'JSXMemberExpression') {
-    return `${jsxNode.name.object.name}.${jsxNode.name.property.name}`
   }
 }
